refactor : 리포지토리 계층의 서비스 의존 제거
리포지토리가 service.dto.PopularTheme을 import하면서
의존 방향이 거꾸로 흐르던 문제를 해결.

- 영속성 단위는 repository 계층에 응집하도록 projection 패키지로 모음.
- PopularTheme은 집계 결과라 Entity로 부르기 어색해 Projection으로 이름 수정.
