+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -4,6 +4,7 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import roomescape.domain.ReservationTime;
+import roomescape.repository.projection.ReservationTimeEntity;
 
 public interface ReservationTimeRepository {
     List<ReservationTimeEntity> findAll();
