+++ src/main/java/roomescape/repository/ThemeRepository.java
@@ -3,16 +3,17 @@ package roomescape.repository;
 import java.util.List;
 import java.util.Optional;
 import roomescape.domain.Theme;
-import roomescape.service.dto.PopularTheme;
+import roomescape.repository.projection.PopularThemeProjection;
+import roomescape.repository.projection.ThemeEntity;
 
 public interface ThemeRepository {
     List<ThemeEntity> findAll();
 
     Optional<ThemeEntity> findById(Long id);
-    
+
     ThemeEntity save(Theme theme);
 
     void deleteById(Long id);
 
-    List<PopularTheme> findPopular();
+    List<PopularThemeProjection> findPopular();
 }
