+++ src/main/java/roomescape/service/ThemeService.java
@@ -3,8 +3,8 @@ package roomescape.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.domain.Theme;
-import roomescape.repository.ThemeEntity;
 import roomescape.repository.ThemeRepository;
+import roomescape.repository.projection.ThemeEntity;
 import roomescape.service.dto.PopularThemeResult;
 import roomescape.service.dto.ThemeCreateCommand;
 import roomescape.service.dto.ThemeResult;
