+++ src/main/java/roomescape/service/dto/ReservationResult.java
@@ -2,7 +2,7 @@ package roomescape.service.dto;
 
 import java.time.LocalDate;
 import roomescape.domain.Reservation;
-import roomescape.repository.ReservationEntity;
+import roomescape.repository.projection.ReservationEntity;
 
 public class ReservationResult {
